use measured queuing delay (as opposed to estimation)

+++ include/ingress.h
@@ -23,7 +23,7 @@ class Ingress {
         std::shared_ptr<RPCMessage> dequeue(std::string);
         size_t size(std::string);
         void update_stats(int32_t, int32_t, std::string&);
-        int64_t add_to_be_admitted_or_drop(RPCQueue&, RPCMapper&, std::string&, int64_t, float, int32_t);
+        int64_t add_to_be_admitted_or_drop(RPCQueue&, RPCMapper&, std::string&, int64_t, int32_t, int32_t);
         void dump_state();
